+++ .gitignore
@@ -1,3 +1,6 @@
 /target
 /dist
-/src/styles.css
+node_modules
+package-lock.json
+npm-debug.log*
+/public/tailwind.css
