+++ Cargo.toml
@@ -12,6 +12,7 @@ cable = { git = "https://github.com/ahdinosaur/cable.rs", branch = "coven" }
 cable_core = { git = "https://github.com/ahdinosaur/cable.rs", branch = "coven" }
 chrono = "0.4.31"
 dioxus = "0.4.0"
+dioxus-daisyui = "0.4.0"
 dioxus-desktop = "0.4.0"
 dioxus-router = "0.4.1"
 dirs = "5.0.1"
@@ -19,11 +20,3 @@ fermi = "0.4.0"
 futures-util = "0.3.29"
 nest = "1.0.0"
 tokio = "1.33.0"
-turf = "0.7.0"
-
-[package.metadata.turf]
-minify = false
-load_paths = ["src/styles"]
-
-[package.metadata.turf.file_output]
-global_css_file_path = "src/styles.css"
