+++ Dioxus.toml
@@ -23,13 +23,13 @@ title = "Coven | p2p chat"
 
 index_on_404 = true
 
-watch_path = ["src"]
+watch_path = ["src", "public"]
 
 # include `assets` in web platform
 [web.resource]
 
 # CSS style file
-style = []
+style = ["/tailwind.css"]
 
 # Javascript code file
 script = []
