+++ src/components/app_nav.rs
@@ -1,0 +1,3 @@
+
+    let cabal_ids = use_read(cx, &CABAL_IDS);
+    let _active_cabal_id = use_read(cx, &ACTIVE_CABAL_ID);
