--- src/components/cabal_container.scss
@@ -1,18 +1,0 @@
-.component {
-  display: flex;
-  flex-direction: column;
-
-  > .top {
-  }
-
-  > .bottom {
-    display: flex;
-    flex-direction: row;
-
-    > .side {
-    }
-
-    > .content {
-    }
-  }
-}
