+++ src/components/mod.rs
@@ -1,13 +1,7 @@
-pub(crate) mod app_container;
-pub(crate) mod app_top_nav;
-pub(crate) mod cabal_container;
-pub(crate) mod cabals_nav;
+pub(crate) mod app_side_nav;
 pub(crate) mod channel_posts;
 pub(crate) mod channels_nav;
 
-pub(crate) use self::app_container::AppContainer;
-pub(crate) use self::app_top_nav::AppTopNav;
-pub(crate) use self::cabal_container::CabalContainer;
-pub(crate) use self::cabals_nav::CabalsNav;
+pub(crate) use self::app_side_nav::AppSideNav;
 pub(crate) use self::channel_posts::ChannelPosts;
 pub(crate) use self::channels_nav::ChannelsNav;
